requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
databases==0.6.2
fastapi==0.88.0
fastapi-pagination==0.11.0
greenlet==2.0.1
h11==0.14.0
httptools==0.5.0
idna==3.4
numpy==1.23.5
Paginator==0.5.1
pandas==1.5.2
pydantic==1.10.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.6
PyYAML==6.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4